requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
azure.core
python-dotenv==1.0.0
aiohttp
azure.ai.textanalytics